nrf52: rename radio to ble_radio
