+++ chips/nrf52/src/chip.rs
@@ -1,8 +1,8 @@
 use crate::adc;
+use crate::ble_radio;
 use crate::deferred_call_tasks::DeferredCallTask;
 use crate::i2c;
 use crate::nvmc;
-use crate::radio;
 use crate::spi;
 use crate::uart;
 use cortexm4::{self, nvic};
@@ -56,7 +56,7 @@ impl kernel::Chip for NRF52 {
                     match interrupt {
                         peripheral_interrupts::ECB => nrf5x::aes::AESECB.handle_interrupt(),
                         peripheral_interrupts::GPIOTE => nrf5x::gpio::PORT.handle_interrupt(),
-                        peripheral_interrupts::RADIO => radio::RADIO.handle_interrupt(),
+                        peripheral_interrupts::RADIO => ble_radio::RADIO.handle_interrupt(),
                         peripheral_interrupts::RNG => nrf5x::trng::TRNG.handle_interrupt(),
                         peripheral_interrupts::RTC1 => nrf5x::rtc::RTC.handle_interrupt(),
                         peripheral_interrupts::TEMP => nrf5x::temperature::TEMP.handle_interrupt(),
